Allow users to delete jobs
