+++ test/acceptance/job_test.exs
@@ -13,4 +13,14 @@ defmodule Community.Acceptance.JobTest do
     assert flash_text =~ "Job updated"
     assert visible_page_text =~ "New Title"
   end
+
+  test "deleting a job" do
+    job = create(:job)
+    navigate_to job_path(@endpoint, :edit, job, token: job.token)
+
+    ignore_confirm_dialog
+    click_on "Delete Job"
+
+    assert flash_text =~ "Job deleted"
+  end
 end
