+++ web/templates/layout/app.html.eex
@@ -21,10 +21,7 @@
         <%= page_action_link(assigns) %>
       </header>
 
-      <div data-role="flash">
-        <p class="alert alert-info" role="alert"><%= get_flash(@conn, :info) %></p>
-        <p class="alert alert-danger" role="alert"><%= get_flash(@conn, :error) %></p>
-      </div>
+      <%= render Community.SharedView, "flashes.html", conn: @conn %>
 
       <main role="main">
         <%= render @view_module, @view_template, assigns %>
