+++ web/templates/layout/landing.html.eex
@@ -27,8 +27,7 @@
         </h2>
       </header>
 
-      <p class="alert alert-info" role="alert"><%= get_flash(@conn, :info) %></p>
-      <p class="alert alert-danger" role="alert"><%= get_flash(@conn, :error) %></p>
+      <%= render Community.SharedView, "flashes.html", conn: @conn %>
 
       <main role="main">
         <%= render @view_module, @view_template, assigns %>
